+++ Tests/SentryTests/Integrations/Session/SentrySessionTrackerTests.swift
@@ -60,25 +60,25 @@ class SentrySessionTrackerTests: XCTestCase {
         super.setUp()
         
         fixture = Fixture()
-        
+
         SentryDependencyContainer.sharedInstance().dateProvider = fixture.currentDateProvider
 
         fixture.fileManager.deleteCurrentSession()
         fixture.fileManager.deleteCrashedSession()
         fixture.fileManager.deleteTimestampLastInForeground()
-        
+
         fixture.setNewHubToSDK()
-        
+
         sut = fixture.getSut()
     }
     
     override func tearDown() {
         stopSut()
         clearTestState()
-        
+
         super.tearDown()
     }
-    
+
     func testOnlyForeground() throws {
         // -- Arrange --
         startSutInAppDelegate()
@@ -196,8 +196,18 @@ class SentrySessionTrackerTests: XCTestCase {
     
     func testCrashInBackground_LaunchInForeground() throws {
         // -- Arrange --
-        crashInBackground()
-        assertNoSessionSent()
+        // During testing we observed that the deallocation of the session tracker happens after the method returns
+        // and not immediately when a new `sut` is set.
+        // This causes multiple session tracker to be registered as observers, until they are fully released and
+        // weak references are nil.
+        // Using an autoreleasepool to ensure that the deallocation happens before the test continues.
+        autoreleasepool {
+            crashInBackground()
+            assertNoSessionSent()
+
+            // Manually deallocate the previous sut to avoid race-conditions of duplicate observers
+            sut = nil
+        }
 
         // -- Act & Assert --
         sut = fixture.getSut()
@@ -343,26 +353,37 @@ class SentrySessionTrackerTests: XCTestCase {
         // -- Assert --
         assertEndSessionSent(started: sessionStarted, duration: 1)
     }
-    
+
     func testForeground_Background_Terminate_LaunchAgain() throws {
+        // During testing we observed that the deallocation of the session tracker happens after the method returns
+        // and not immediately when a new `sut` is set.
+        // This causes multiple session tracker to be registered as observers, until they are fully released and
+        // weak references are nil.
+        // Using an autoreleasepool to ensure that the deallocation happens before the test continues.
         // -- Arrange --
         let sessionStartTime = fixture.currentDateProvider.date()
         startSutInAppDelegate()
 
         // -- Act --
-        goToForeground()
-        advanceTime(bySeconds: 1)
-        goToBackground()
-        
-        advanceTime(bySeconds: 10)
-        terminateApp()
-        assertEndSessionSent(started: sessionStartTime, duration: 1)
-        stopSut()
+        autoreleasepool {
+            goToForeground()
+            advanceTime(bySeconds: 1)
+            goToBackground()
 
-        advanceTime(bySeconds: 1)
-        
-        // Launch the app again
-        fixture.setNewHubToSDK()
+            advanceTime(bySeconds: 10)
+            terminateApp()
+            assertEndSessionSent(started: sessionStartTime, duration: 1)
+            stopSut()
+
+            advanceTime(bySeconds: 1)
+
+            // Launch the app again
+            fixture.setNewHubToSDK()
+
+            // Manually deallocate the previous sut to avoid race-conditions of duplicate observers
+            sut = nil
+        }
+
         sut = fixture.getSut()
 
         startSutInAppDelegate()
@@ -438,6 +459,50 @@ class SentrySessionTrackerTests: XCTestCase {
         assertNotificationNames(notificationNames)
     }
 
+    func testForegroundBeforeStart_shoudStartSession() throws {
+        // -- Arrange --
+        goToForeground()
+
+        // Pre-condition: No session should be sent yet
+        assertNoSessionSent()
+
+        // -- Act --
+        sut.start()
+
+        // -- Assert --
+        XCTAssertEqual(fixture.client.captureSessionInvocations.invocations.count, 1)
+        let startSessionEvent = try XCTUnwrap(fixture.client.captureSessionInvocations.invocations.element(at: 0))
+        XCTAssertEqual(startSessionEvent.status, SentrySessionStatus.ok)
+    }
+
+    func testRestartInForeground_shouldStartNewSession() throws {
+        // -- Arrange --
+        var startTime = fixture.currentDateProvider.date()
+        startSutInAppDelegate()
+        goToForeground()
+
+        sut.stop()
+        assertSessionsSent(count: 1)
+        try assertSessionInitSent(sessionStarted: startTime)
+
+        // -- Act --
+        startTime = fixture.currentDateProvider.date()
+        sut.start()
+
+        // -- Assert --
+        // We expect 3 session events of two sessions:
+        // 1. The init session that was sent when the app started
+        // 2. The end session that was sent when the app stopped
+        // 3. The init session that was sent when the app restarted
+        XCTAssertEqual(fixture.client.captureSessionInvocations.invocations.count, 3)
+        let startSessionEvent = try XCTUnwrap(fixture.client.captureSessionInvocations.invocations.element(at: 0))
+        XCTAssertEqual(startSessionEvent.status, SentrySessionStatus.ok)
+        let endSessionEvent = try XCTUnwrap(fixture.client.captureSessionInvocations.invocations.element(at: 1))
+        XCTAssertEqual(endSessionEvent.status, SentrySessionStatus.exited)
+        let restartSessionEvent = try XCTUnwrap(fixture.client.captureSessionInvocations.invocations.element(at: 2))
+        XCTAssertEqual(restartSessionEvent.status, SentrySessionStatus.ok)
+    }
+
     // MARK: - Helpers
 
     private func startSutInAppDelegate() {
@@ -481,7 +546,14 @@ class SentrySessionTrackerTests: XCTestCase {
         // This can be observed by viewing the application state in `UIAppDelegate.applicationDidBecomeActive`.
         fixture.application.applicationState = .active
         #endif
-        Dynamic(sut).didBecomeActive()
+        fixture.notificationCenter
+            .post(
+                Notification(
+                    name: SentryNSNotificationCenterWrapper.didBecomeActiveNotificationName,
+                    object: nil,
+                    userInfo: nil
+                )
+            )
     }
     
     private func goToBackground() {
@@ -491,7 +563,14 @@ class SentrySessionTrackerTests: XCTestCase {
         // It is expected that the app state is background when the didEnterBackground is called
         fixture.application.applicationState = .background
         #endif
-        Dynamic(sut).didEnterBackground()
+        fixture.notificationCenter
+           .post(
+               Notification(
+                   name: SentryNSNotificationCenterWrapper.didEnterBackgroundNotificationName,
+                   object: nil,
+                   userInfo: nil
+               )
+           )
     }
     
     private func willResignActive() {
@@ -500,7 +579,14 @@ class SentrySessionTrackerTests: XCTestCase {
         // This can be observed by viewing the application state in `UIAppDelegate.applicationWillResignActive`.
         fixture.application.applicationState = .active
         #endif
-        Dynamic(sut).willResignActive()
+        fixture.notificationCenter
+            .post(
+                Notification(
+                    name: SentryNSNotificationCenterWrapper.willResignActiveNotificationName,
+                    object: nil,
+                    userInfo: nil
+                )
+            )
     }
     
     private func hybridSdkDidBecomeActive() {
@@ -508,7 +594,14 @@ class SentrySessionTrackerTests: XCTestCase {
         // When an app did become active, it is in the active state.
         fixture.application.applicationState = .active
         #endif
-        Dynamic(sut).didBecomeActive()
+        fixture.notificationCenter
+            .post(
+                Notification(
+                    name: SentryNSNotificationCenterWrapper.didBecomeActiveNotificationName,
+                    object: nil,
+                    userInfo: nil
+                )
+            )
     }
     
     private func goToBackground(forSeconds: TimeInterval) {
@@ -523,7 +616,14 @@ class SentrySessionTrackerTests: XCTestCase {
         #if os(iOS) || targetEnvironment(macCatalyst) || os(tvOS)
         fixture.application.applicationState = .background
         #endif
-        Dynamic(sut).willTerminate()
+        fixture.notificationCenter
+            .post(
+                Notification(
+                    name: SentryNSNotificationCenterWrapper.willTerminateNotificationName,
+                    object: nil,
+                    userInfo: nil
+                )
+            )
     }
     
     private func terminateApp() {
@@ -701,6 +801,10 @@ class SentrySessionTrackerTests: XCTestCase {
         override var applicationState: UIApplication.State {
             get { _underlyingAppState }
             set { _underlyingAppState = newValue }
+        }
+
+        override func isActive() -> Bool {
+            return applicationState == .active
         }
     }
 #endif